feat: CustomProgressView 생성 및 InfoStep 수정
